Add files via upload
This project is aimed at developing a Android Application that depicts online News .The DigiNewspaper is an Advanced and Informative System helping the users to know the news, articles or any interesting things happening all over the world. 
This App helps the local public to keep themselves up to date about the happenings around their vicinity or locality or in their City.